requirements.txt
beautifulsoup4
dnspython
folium
fpdf
geoip2
geopy
langchain
langchain_experimental
langchain_google_genai
matplotlib
networkx
numpy
pandas
plotly
protobuf
python-dotenv
pythonwhois==2.2.2
regex
requests
scikit_learn
seaborn
streamlit
streamlit_lottie
streamlit_option_menu
